+++ declarations/ninjagoat.d.ts
@@ -21,7 +21,7 @@ export let lazyInject: (serviceIdentifier: string | symbol | interfaces.Newable<
 export let lazyMultiInject: (serviceIdentifier: string | symbol | interfaces.Newable<any> | interfaces.Abstract<any>) => (proto: any, key: string) => void;
 
 export interface Dictionary<T> {
-    [index: string]: T
+    [index: string]: T;
 }
 
 export interface IModule {
@@ -57,6 +57,8 @@ export interface IViewModelRegistry {
     index<T>(construct: interfaces.Newable<IViewModel<T>>, observable?: (context: ViewModelContext) => Rx.IObservable<T>): AreaRegistry;
     notFound<T>(construct: interfaces.Newable<IViewModel<T>>, observable?: (context: ViewModelContext) => Rx.IObservable<T>): AreaRegistry;
     add<T>(construct: interfaces.Newable<IViewModel<T>>, observable?: (context: ViewModelContext) => Rx.IObservable<T>, parameters?: string): IViewModelRegistry;
+    withParameters(parameters: string): IViewModelRegistry;
+    notifyBy(notify: (parameters: any) => string): IViewModelRegistry;
     forArea(area: string): AreaRegistry;
     getArea(areaId: string): AreaRegistry;
     getAreas(): AreaRegistry[];
@@ -70,15 +72,15 @@ export class AreaRegistry {
 }
 
 export class RegistryEntry<T> {
+
     construct: interfaces.Newable<IViewModel<T>>;
     id: string;
     observableFactory: (context: ViewModelContext) => Rx.IObservable<T>;
-    parameters: string;
+    parameters?: string;
+    notify?: (parameters: any) => string;
 
-    constructor(construct: interfaces.Newable<IViewModel<T>>,
-                id: string,
-                observableFactory: (context: ViewModelContext) => Rx.IObservable<T>,
-                parameters: string);
+    constructor(construct: interfaces.Newable<IViewModel<T>>, id: string,
+                observableFactory: (context: ViewModelContext) => Rx.IObservable<T>, parameters?: string);
 }
 
 export class ViewModelContext {
@@ -90,16 +92,20 @@ export class ViewModelContext {
 }
 
 export interface IViewModelFactory {
-    create<T extends IViewModel<T>>(context: ViewModelContext, construct: interfaces.Newable<IViewModel<T>>,
-                                    observableFactory: (context: ViewModelContext) => Rx.IObservable<T>): T;
+    create<T>(context: ViewModelContext, construct: interfaces.Newable<IViewModel<T>>,
+              observableFactory: (context: ViewModelContext) => Rx.Observable<T>): IViewModel<T>;
+}
+
+export interface IViewModelFactoryExtender {
+    extend<T>(viewmodel: T, context: ViewModelContext, source: Rx.Observable<T>);
 }
 
 export class ViewModelFactory implements IViewModelFactory {
 
-    constructor(container: IObjectContainer);
+    constructor(container: IObjectContainer, extenders: IViewModelFactoryExtender[]);
 
-    create<T extends IViewModel<T>>(context: ViewModelContext, construct: interfaces.Newable<IViewModel<T>>,
-                                    observableFactory: (context: ViewModelContext) => Rx.IObservable<T>): T;
+    create<T>(context: ViewModelContext, construct: interfaces.Newable<IViewModel<T>>,
+              observableFactory: (context: ViewModelContext) => Rx.Observable<T>): IViewModel<T>;
 }
 
 export interface IViewModel<T> extends Rx.IDisposable, Rx.IObservable<void> {
@@ -263,7 +269,7 @@ export interface IRoutingAdapter {
 export function FeatureToggle(predicate: CheckPredicate);
 
 export interface CheckPredicate {
-    (): boolean
+    (): boolean;
 }
 
 export interface IFeatureChecker {