+++ scripts/registry/IViewModelRegistry.ts
@@ -5,11 +5,15 @@ import ViewModelContext from "../registry/ViewModelContext";
 import {interfaces} from "inversify";
 import * as Rx from "rx";
 
+
+
 interface IViewModelRegistry {
     master<T>(construct: interfaces.Newable<IViewModel<T>>, observable?: (context: ViewModelContext) => Rx.IObservable<T>): AreaRegistry;
     index<T>(construct: interfaces.Newable<IViewModel<T>>, observable?: (context: ViewModelContext) => Rx.IObservable<T>): AreaRegistry;
     notFound<T>(construct: interfaces.Newable<IViewModel<T>>, observable?: (context: ViewModelContext) => Rx.IObservable<T>): AreaRegistry;
     add<T>(construct: interfaces.Newable<IViewModel<T>>, observable?: (context: ViewModelContext) => Rx.IObservable<T>, parameters?: string): IViewModelRegistry;
+    withParameters(parameters: string): IViewModelRegistry;
+    notifyBy(notify: (parameters: any) => string): IViewModelRegistry;
     forArea(area: string): AreaRegistry;
     getArea(areaId: string): AreaRegistry;
     getAreas(): AreaRegistry[];